+++ CHANGELOG.md
@@ -9,6 +9,7 @@ ChangeLog
 * Fixed a privilege bug in the availability system.
 * #697: Added a "tableName" member to the PropertyStorage PDO backend. (@Frzk).
 * #699: PostgreSQL fix for the Locks PDO backend. (@TCKnet)
+* Removed the `simplefsserver.php` example file. It's not simple enough.
 
 
 3.1.0-alpha1 (2015-07-19)
@@ -27,6 +28,8 @@ ChangeLog
 ------------------
 
 * #703: PropPatch in client is not correctly encoded.
+* #709: Throw exception when running into empty
+  `supported-calendar-component-set.
 
 
 3.0.3 (2015-08-06)